Add max health for base entities (#35)
* Add max health for base entities

I added max health for the base entities. For example, if the max HP of a player is 100, the player's health should be no more than 100. 

I also added a method that applies damage to a base entity. I added this to the base entity class. I added it to that base entity because I thought that every base entity will have a max health. I think that we should use a method, as opposed to changing the attribute directly because this will allow us to add check when change the health points.

Need to update the lines of code that used called the base entity class

* Add max health indicator to Heads-up display (HUD)