Map selector (#36)
* Add max health for base entities

I added max health for the base entities. For example, if the max HP of a player is 100, the player's health should be no more than 100. 

I also added a method that applies damage to a base entity. I added this to the base entity class. I added it to that base entity because I thought that every base entity will have a max health. I think that we should use a method, as opposed to changing the attribute directly because this will allow us to add check when change the health points.

Need to update the lines of code that used called the base entity class

* Added an option to select map

* Added the ability to choose the map you want to load

* Updated Mateo01 map to work

* Updated "NEw World" map

- I updated the "New World" map so that it worked with current program
- Also, I also squashed a bug where the map size indicated in the JSON file was not what was being displayed.
- Also, added code for the Void Tile in the RenderingManager (just in case someone wants to use it)

* Forgot to save and commit